+++ processing/pipeline.py
@@ -5,17 +5,17 @@ import time
 from queue import Queue
 import sys
 
-from node import Node
-from detect_connections import detect_connections
-from detect_nodes import detect_nodes
-from template_library import TemplateLibrary
-from drawer import draw_map
-from grabber import switch_window, screenshot, do_drag_move, move_mouse, mock_switch_window, mock_move_screen, \
-    mock_screenshot, DragListener, MockDragListener
-from pathfinder import run_pathfinder
-from process_map import Finalizer
-from path_converter import get_path
-from settings import Settings
+from data.node import Node
+from processing.detect_connections import detect_connections
+from processing.detect_nodes import detect_nodes
+from processing.template_library import TemplateLibrary
+from front.drawer import draw_map
+from front.grabber import switch_window, screenshot, do_drag_move, move_mouse, mock_switch_window, mock_move_screen, \
+    mock_screenshot, DragListener
+from processing.pathfinder import run_pathfinder
+from processing.process_map import Finalizer
+from utils.path_converter import get_path
+from data.settings import Settings
 
 """
 Full process (may be outdated a bit):
