+++ processing/process_map.py
@@ -2,8 +2,8 @@
 import json
 import os
 
-from detect_connections import detect_connections
-from path_converter import get_path
+from processing.detect_connections import detect_connections
+from utils.path_converter import get_path
 
 MAX_VERTICAL_ROWS = 5
 HORIZONTAL_OFFSETS = [0, -1, -2, -3, -4, -5, -6]
