+++ utils/build.py
@@ -7,8 +7,10 @@ from pathlib import Path
 import fnmatch
 import sys
 
-from settings import Settings
-from score_config import CURRENT_SEASON
+from data.settings import Settings
+from data.score_config import CURRENT_SEASON
+
+BASE_DIR = Path(__file__).resolve().parent.parent
 
 SOURCE_SCRIPT = "gui.py"
 EXE_NAME = "CZN Pathfinder"
@@ -44,20 +46,20 @@ release_ignore = [
 
 ]
 
-DIST_DIR = Path("dist")
-BUILD_DIR = Path("build")
-FINAL_OUTPUT_DIR = Path("Exe_build_folder")
+DIST_DIR = BASE_DIR / "dist"
+BUILD_DIR = BASE_DIR / "build"
+FINAL_OUTPUT_DIR = BASE_DIR / "Exe_build_folder"
 
 
 def insert_custom_backgrounds(expected):
-    path = Path("Images/Map_background")
+    path = BASE_DIR / "Images" / "Map_background"
     for image in os.listdir(path):
         if image.endswith("_g.png"):
             path_t = path / image
             expected.append(path_t)
 
 def load_gitignore():
-    path = Path(".gitignore")
+    path = BASE_DIR / ".gitignore"
     if not path.exists():
         return []
     rules = []
@@ -119,7 +121,7 @@ def match_gitignore(path: Path, rules):
 
 
 def prepare():
-    root = Path(".").resolve()
+    root = BASE_DIR.resolve()
     gitignore_rules = load_gitignore()
     result = []
 
@@ -160,12 +162,12 @@ def prepare():
 
 def expand_expected_items(items):
     expanded = []
-    root = Path(".")
+    root = BASE_DIR
 
     for item in items:
-        p = Path(item)
+        p = BASE_DIR / item
         if not p.exists():
-            raise RuntimeError(f"expected_item not found: {item}")
+            raise RuntimeError(f"expected_item not found: {item} at path {p}")
 
         if p.is_file():
             expanded.append(p.relative_to(root).as_posix())
@@ -208,9 +210,10 @@ def clean_previous_builds():
             shutil.rmtree(d)
     to_be_removed = []
 
-    for f in os.listdir("Example_scan_result"):
+    example_dir = BASE_DIR / "Example_scan_result"
+    for f in os.listdir(example_dir):
         if not f.startswith("map"):
-            to_be_removed.append(os.path.join("Example_scan_result", f))
+            to_be_removed.append(example_dir / f)
     for f in to_be_removed:
         os.remove(f)
 
@@ -221,7 +224,7 @@ def build_executable():
         "--clean",
         "CZN Pathfinder.spec"
     ]
-    subprocess.run(cmd, check=True)
+    subprocess.run(cmd, check=True, cwd=BASE_DIR)
 
 
 def create_zip():
@@ -229,7 +232,7 @@ def create_zip():
     zip_path = FINAL_OUTPUT_DIR / ZIP_NAME
     expanded = expand_expected_items(expected_items)
     exe_file = EXE_NAME + ".exe"
-    exe_path = Path("dist") / exe_file
+    exe_path = DIST_DIR / exe_file
     if not exe_path.exists():
         raise FileNotFoundError("No exe found")
 
@@ -237,7 +240,7 @@ def create_zip():
     with zipfile.ZipFile(zip_path, "w", zipfile.ZIP_DEFLATED) as z:
         z.write(str(exe_path), folder_prefix + Path(exe_path).name)
         for item in expanded:
-            z.write(item, folder_prefix + item)
+            z.write(BASE_DIR / item, folder_prefix + item)
 
 
 def main():
